requirements.txt
fastapi
uvicorn
faiss-cpu
numpy
requests
sentence-transformers
google-generativeai
python-dotenv